processing/requirements-clean.txt
acres==0.3.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs
beautifulsoup4
Brotli
cachetools==5.5.2
certifi
cffi
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudinary==1.42.2
configobj==5.0.9
configparser==7.1.0
deprecation==2.1.0
dotenv==0.9.9
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup
filelock==3.17.0
fitz==0.0.1.dev2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
gotrue==2.11.4
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11
h2
hpack
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe
idna
isodate==0.6.1
jsonpath-python==1.0.6
looseversion==1.3.0
lxml==5.3.1
mistralai==1.5.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
outcome
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.1.0
postgrest==0.19.3
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
PySocks
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyxnat==1.6.3
rdflib==6.3.2
realtime==2.4.1
requests==2.32.3
rsa==4.9
scipy==1.15.2
selenium
simplejson==3.20.1
six==1.17.0
sniffio
sortedcontainers
soupsieve
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tqdm==4.67.1
traits==7.0.2
trio
trio-websocket
typing-inspect==0.9.0
typing_extensions
tzdata==2025.1
uritemplate==4.1.1
urllib3
websocket-client
websockets==14.2
wsproto
yarl==1.18.3
zstandard==0.23.0
boto3
psycopg2-binary
together
